+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Removed support for Python 3.6, 3.7 ([#23])
 * Added support for Python 3.10, 3.11 ([#23])
 
+### Fixed
+
+* Parse errors on loading a grammar are now GrammarError ([#21])
+
 
 ## [v0.3.2][]
 
@@ -132,4 +136,5 @@ descent parser and a work-in-progress state-machine parser.
 [#18]: https://github.com/goodmami/pe/issues/18
 [#19]: https://github.com/goodmami/pe/issues/19
 [#20]: https://github.com/goodmami/pe/issues/20
+[#21]: https://github.com/goodmami/pe/issues/21
 [#23]: https://github.com/goodmami/pe/issues/23